+++ Scenes/Modules/Interactables/DialogueInteractable.tscn
@@ -1,0 +1,8 @@
+[gd_scene load_steps=2 format=3 uid="uid://ouc1bluwl020"]
+
+[ext_resource type="Script" path="res://Scripts/Modules/Interactables/DialogueInteractable.cs" id="1_q7tdu"]
+
+[node name="DialogueInteractable" type="Area2D"]
+collision_layer = 8
+collision_mask = 8
+script = ExtResource("1_q7tdu")
