+++ Scripts/Modules/Interactables/DialogueInteractable.cs
@@ -5,21 +5,22 @@ using TheLoneLanternProject.Scripts.Shared.Constants;
 using TheLoneLanternProject.Scripts.Shared.Helpers;
 using TheLoneLanternProject.Scripts.Utils.Signals;
 
-namespace TheLoneLanternProject.Scripts.Modules.Interactables;
 
-[GlobalClass]
-public partial class DialogueInteractableModule : Node
+
+public partial class DialogueInteractableModule : Area2D, IInteractable
 {
+    [Export] public DialogueManager dialogueScript; // fix this 
+
     private CustomSignals customSignals = new();
     private Luce luce = new();
     public override void _Ready()
     {
         customSignals = GetNode<CustomSignals>("/root/CustomSignals");
         customSignals.Interaction += DoSomething;
-        customSignals.ShowDialogueBalloon += ShowDialogueBalloon;
+        /*customSignals.ShowDialogueBalloon += Interact;
 
         var tree = GetTree();
-        luce = GetNodeHelper.GetLuce(tree);
+        luce = GetNodeHelper.GetLuce(tree);*/
     }
 
     private void DoSomething()
@@ -27,17 +28,17 @@ public partial class DialogueInteractableModule : Node
         GD.Print("Picked Up!");
     }
 
-    private void ShowDialogueBalloon(string dialogue, string title)
+    public void Interact()
     {
-        luce.SetState(PlayerState.Disabled);
+        /*luce.SetState(PlayerState.Disabled);
 
         DialogueManager.ShowDialogueBalloon(GD.Load($"res://Assets/Dialogue/{dialogue}.dialogue"), title);
-        DialogueManager.DialogueEnded += SetupGameplayAfterDialogueEnded;
+        DialogueManager.DialogueEnded += SetupGameplayAfterDialogueEnded;*/
     }
 
-    private void SetupGameplayAfterDialogueEnded(Resource dialogueResource)
+    /*private void SetupGameplayAfterDialogueEnded(Resource dialogueResource)
     {
         luce.SetState(PlayerState.Idle);
         DialogueManager.DialogueEnded -= SetupGameplayAfterDialogueEnded;
-    }
+    }*/
 }
