--- Scripts/Modules/Interactables/PickupInteractableModule.cs
@@ -1,21 +1,0 @@
-using Godot;
-
-using TheLoneLanternProject.Scripts.Utils.Signals;
-
-namespace TheLoneLanternProject.Scripts.Modules.Interactables; // May not be needed
-
-[GlobalClass]
-public partial class PickupInteractableModule : Node
-{
-    private CustomSignals customSignals = new();
-    public override void _Ready()
-    {
-        customSignals = GetNode<CustomSignals>("/root/CustomSignals");
-        customSignals.Interaction += DoSomething;
-    }
-
-    private void DoSomething()
-    {
-        GD.Print("Picked Up!");
-    }
-}
