+++ Scripts/Modules/Interactables/PlayerInteractorModule.cs
@@ -1,0 +1,7 @@
+using Godot;
+using System;
+
+[GlobalClass]
+public partial class PlayerInteractorModule : Node
+{
+}
